add builder to pojo classes Signed-off-by: dbulahov <bulahovdenis@gmail.com>
