+++ structure.emf/test.bndrun
@@ -70,7 +70,6 @@
 	org.eclipse.emf.ecore.xmi;version='[2.37.0,2.37.1)',\
 	org.gecko.emf.osgi.component;version='[6.1.3,6.1.4)',\
 	org.opentest4j;version='[1.3.0,1.3.1)',\
-	org.osgi.service.cm;version='[1.6.1,1.6.2)',\
 	org.osgi.service.component;version='[1.5.1,1.5.2)',\
 	org.osgi.test.common;version='[1.3.0,1.3.1)',\
 	org.osgi.test.junit5;version='[1.3.0,1.3.1)',\
