+++ structure.pojo/src/main/java/org/eclipse/daanse/rdb/structure/pojo/ViewTableImpl.java
@@ -13,15 +13,18 @@
  */
 package org.eclipse.daanse.rdb.structure.pojo;
 
+import java.util.List;
+
 import org.eclipse.daanse.rdb.structure.api.model.ViewTable;
+import org.eclipse.daanse.rdb.structure.pojo.InlineTableImpl.Builder;
 
 public class ViewTableImpl extends AbstractTable implements ViewTable {
 
     public ViewTableImpl(Builder builder) {
-        setName(builder.getName());
-        setColumns(builder.getColumns());
-        setSchema(builder.getSchema());
-        setDescription(builder.getDescription());
+        setName(builder.name);
+        setColumns(builder.columns);
+        setSchema(builder.schema);
+        setDescription(builder.description);
     }
 
     public static Builder builder() {
@@ -31,6 +34,26 @@ public class ViewTableImpl extends AbstractTable implements ViewTable {
     public static final class Builder extends AbstractBuilder {
 
         private Builder() {
+        }
+
+        public Builder withName(String name) {
+            this.name = name;
+            return this;
+        }
+
+        public Builder withColumns(List<ColumnImpl> columns) {
+            this.columns = columns;
+            return this;
+        }
+
+        public Builder withsSchema(DatabaseSchemaImpl schema) {
+            this.schema = schema;
+            return this;
+        }
+
+        public Builder withsDdescription(String description) {
+            this.description = description;
+            return this;
         }
 
         public ViewTableImpl build() {
