+++ structure.reader/test.bndrun
@@ -79,6 +79,7 @@
 	org.gecko.emf.osgi.api;version='[6.1.3,6.1.4)',\
 	org.gecko.emf.osgi.component;version='[6.1.3,6.1.4)',\
 	org.opentest4j;version='[1.3.0,1.3.1)',\
+	org.osgi.service.cm;version='[1.6.1,1.6.2)',\
 	org.osgi.service.component;version='[1.5.1,1.5.2)',\
 	org.osgi.test.common;version='[1.3.0,1.3.1)',\
 	org.osgi.test.junit5;version='[1.3.0,1.3.1)',\
